Added buttons to navigate through questions

+++ src/components/CVSectionQuestions/SkillsQuestions.js
@@ -9,18 +9,22 @@ class SkillsQuestions extends Component {
 
 
     render() {
-        console.log(this.props)
-        return (
+        let skillsQuestions
+        if (this.props.questionState === 5){
+            skillsQuestions = 
             <div>
-                <div>Skills</div>{}
+            <div>Skills</div>
 
-                {this.props.skills.map((skill, index) => {
-                    return <Input key={index} handleChange={(e) => this.props.captureSkill(e, index)} placeholder="Add Skill"/>
-                })}
-                
-                <Button onSubmit={(e) => this.props.addSkill(e)} text="Add Skill"/>
-                <Button onSubmit={(e) => this.props.deleteSkill(e)} text="Delete Skill"/>
-            </div>
+            {this.props.skills.map((skill, index) => {
+                return <Input key={index} handleChange={(e) => this.props.captureSkill(e, index)} placeholder="Add Skill"/>
+            })}
+            
+            <Button onSubmit={(e) => this.props.addSkill(e)} text="Add Skill"/>
+            <Button onSubmit={(e) => this.props.deleteSkill(e)} text="Delete Skill"/>
+        </div>
+        }
+        return (
+           <div>{skillsQuestions}</div>
         )
     }
 }
